requirements.txt
asttokens==2.4.1
backcall==0.2.0
cloudpickle==3.0.0
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
dill==0.3.9
docopt==0.6.2
executing==2.1.0
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
gym==0.21.0
gym-notices==0.0.8
importlib-metadata==4.13.0
importlib-resources==6.4.5
ipython==8.12.3
jedi==0.19.1
jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
-e git+https://github.com/levensons/UVIP.git@cb0aea0dcfe433d22479f9667eb51a14c7dd4b5c#egg=libUVIP
llvmlite==0.41.1
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
packaging==24.1
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
prompt-toolkit==3.0.48
ptyprocess==0.7.0
pure-eval==0.2.3
pygame==2.6.1
pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
rlberry==0.3.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
stable-baselines3==1.6.2
stack-data==0.6.3
sympy==1.13.3
threadpoolctl==3.5.0
torch==2.4.1
tqdm==4.66.5
traitlets==5.14.3
triton==3.0.0
typing-extensions==4.12.2
tzdata==2024.2
wcwidth==0.2.13
zipp==3.20.2
